Regenerate UFOs using vfbLib and format them using ufoLib2.
Some of the features.fea files did require a bit of manual tweaking, though; dunno if that's a bug in the converter or if the vfb sources were in a bad state.

Since subsequent edits will be made using ufoLib2, and it formats the XML files a little differently than vfbLib does, we do a no-op edit with that library so that future diffs are easier to read.